add generation page
